[red-knot] consolidate diagnostic and type-inference tests
